Method call reference: major rewrite
This section of the reference has been oversimplistic for some time (#1018
and #1534) and various rewrites have been attempted (e.g. #1394, #1432).
Here's another attempt!

My approach here is:
* Stop trying to keep it short and concise
* Document what actually happens in the code, step by step

This does result in a long explanation, because we're trying to document
nearly 2400 lines of code in `probe.rs`, but doing otherwise feels
as though we'll continue to run into criticisms of oversimplification.

This rewrite documents the post-arbitrary-self-types v2 situation,
i.e. it assumes https://github.com/rust-lang/rust/pull/135881 has
landed. We should not merge this until or unless that lands.

This PR was inspired by discussion in #1699. If we go ahead with this
approach, #1699 becomes irrelevant. There was also discussion at
https://github.com/rust-lang/cargo/pull/15117#discussion_r1934734824